+++ public/styles.css
@@ -1,5 +1,5 @@
 body {
-  background-color: darkgray;
+  background-color: #e8ecf2;
   font-family: 'Raleway', sans-serif;
 }
