+++ public/surveys/survey.html
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+<!-- <!DOCTYPE html>
 <html>
   <head>
     <meta charset="utf-8">
@@ -7,6 +7,9 @@
     <link rel="shortcut icon" href="/poll.png" type="image/x-icon">
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
     <link href="https://fonts.googleapis.com/css?family=Faster+One|Raleway" rel="stylesheet">
+    <script src="//code.jquery.com/jquery-3.1.0.min.js"></script>
+    <script src="./form.js"></script>
+    <script src="./create-survey.js"></script>
     <title>form</title>
   </head>
   <body>
@@ -18,7 +21,6 @@
 
 
 
-    <script src="//code.jquery.com/jquery-3.1.0.min.js"></script>
-    <script src="./survey.js"></script>
+
   </body>
-</html>
+</html> -->
