+++ server/index.js
@@ -1,6 +1,7 @@
 const bodyParser = require('body-parser');
 const express = require('express');
 const app = express();
+const http = require('http');
 const md5 = require('md5');
 const path = require('path');
 const passport = require('passport');
@@ -8,26 +9,30 @@ const environment = process.env.NODE_ENV || 'development';
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
+
+
 app.use(express.static(path.resolve(__dirname, '..', 'public')));
 
 app.set('port', process.env.PORT || 3000);
 
 app.locals.surveys = [];
 
-app.post('/api/surveys', function (req, res) {
+app.post('/api/survey', function (req, res) {
   const formData = req.body;
   const id = md5(formData);
   const survey = { id, formData }
   app.locals.surveys.push(survey)
-  res.redirect(`/survey.html?surveyID=${id}`)
+  res.send(survey)
 });
 
 app.get('/api/survey/:surveyID', function (req, res) {
-  const id = req.params.surveyID;
-  let currentSurvey = app.locals.surveys.filter(id);
-  console.log(id, 'id');
-  console.log(currentSurvey, 'currentSurvey');
-  res.status(200).send(currentSurvey);
+  console.log(res);
+  // const id = req.params.surveyID;
+  // let currentSurvey = app.locals.surveys.find((survey) => {
+  //   return survey.surveyID === id
+  // })
+  // console.log(currentSurvey);
+  // res.json(currentSurvey)
 })
 
 app.use((req, res, next) => {
@@ -44,7 +49,6 @@ app.use((err, req, res, next) => {
 app.listen(process.env.PORT || 3000, () => {
     console.log('listening');
 });
-
 
 app.use(passport.initialize());
 app.use(passport.session());
